add simulation script
simulation/soc1 is a testbench for the whole soc1.
soc1.sh is the script to run the simulation.
soc1.f is the file list that add to the simulation and compiled,
  Note: alter_mf.v also added in this file.

soc1.wlf file can be viewed: vsim -view soc1.wlf

Because the simulation modelsim is separate from quartus,
so the simulation project need the ram initialize file in
the current work directory, which is good, so that each test
case can have its own memory(code & data). The ram.mif and
vgaram.mif is for the this soc1 simulation.

need more work to add monitored signal and other tcl stuff.

Later, add more test cases under simulation diretory.