Add new simulation test test2_addi_cpu_clk_initial_1
See doc/test1_addi

this case should fail, it proves that if the very first
instruction at pc 0x0 get filled into pipeline twice, it will
affect the result. (in this testbench, cpu_clk init to 1)

The timing of reset_n pull high is critical. It should be a little
behind a clk rising edge.

Not sure if the "reg cpu_clk" is initialized to 0 when runs on real
fpga.

Need to figure out a mechanism to solve this.