add mret and empty pipe singals
add mret instruction and test13_mret

Found a problem. When the mret is decoded in the IFID stage,
the pcF will be filled with mepc. I choose to do this at IFID stage
rather than do it in a later stage such as EX or MEM.

The problem is that the instruction right before mret may not
be finished in time. For example, if this instruction is csrrw which
updates mepc, the mret will still use the old mepc.

Such as the flowing code that tries to return from the trap handler.

<trap>:
00a00313;  --           addi x6 x0 10
341021f3;  --           csrrs x3 mepc x0 (csrr rd mepc)
00418193;  --           addi x3 x3 4
34119073;  --           csrrw x0 mepc x3 (csrw mepc rs)
30200073;  --           mret

To solve this problem, I come up with this empty pipeline
mechanism. It is not flushing the pipeline. It is to wait for the
instructions that already in the pipeline to finish.

There is a signal empty_pipeline_req that will be passed all the
way through pipelines.

empty_pipeline_req-->empty_pipeline_reqE-->empty_pipeline_reqM
-->empty_pipeline_reqW

Then, at the last pipeline stage, W, signal empty_pipeline_ackW.

In the IFID stage, the logic is the following:

assign empty_pipeline_stallF =   empty_pipeline_req
                               & (~empty_pipeline_ackW);

The pcF is stalled until the empty_pipeline_ackW is signaled.