update vga640x480 module
updateram.sh
add a script to update ram.mif and vgaram.mif
no need to compile the whole project

use the copy under the same directory of this script

For example, taking which ram.mif, depends on the file path in the
ram.v. Since we need to change ram for every test, the path in
ram.v is "ram.mif", instead of "../../soc/cpu6/ram.mif".

But for vgaram.mif vgatextram.mif and chrom.mif, don't need to
change the for every test.

now the vga is in graphic mode,
0x10000 is the first line of the screen, 0x1009f is the last spot.
because 640 / 4 = 160

then 0x10100 is the second line and so forth.

0x10000 is the space of vgaram

---------------------

now the ram.mif draw three blue pixel on the fifth line
then enconter an illegal instruction
the trap handler print a red pixel on the sixth line