Add lic local interrupt controller
Now it only has the timer that run at the same frequency as the
core does.

Only mtime. It does not trigger interrupt yet.

The device MMIO range is 0x20000 - 0x2FFFF

mtime     0x20000
mtimecmp  0x20004

test15_read_mtime tests reading mtime at 0x20000.
Read it twice.

Need triggers interrupt next, and fill mcause.

The testbench is slightly different than running on the FPGA.
It may because of the first clock cycle on mtime register.
The dnext and qout. For now it is ok.