Implment timer; add test19_timer
One thing to notice, when lic_timer_interrupt signaled,
so does excp_flush_pc_ena also, then the pcnextF points to 0x4,
the trap vector. Because when the timer bursts, the current
instruction is the beq, therefore the stallF is also signaled.
pcreg works on the clock falling edge, but the stallF prevent it
from updating the trap vector.

The excp_flush_pc_ena should override the harzard_stallF and
empty_pipeline_stallF. In this case, it is the harazrd_stallF
takes effect here because of the branch instruction.

assign stallF = (~excp_flush_pc_ena &
       	      		(hazard_stallF | empty_pipeline_stallF));

I should document the hazardcontrol module.