Add auipc instruction
aupic needs imm+pc

It is not easy to get the correct pc value because the way I
handle pcF pcE and branch instruction makes it complicated.

pcE changes at the falling edge, because the pcnextE value needs
to be send out before the next raising edge, so that the pcF can
be fetched in time.

working on the falling edge is not a good idea.

for auipc instruction, add pc_auipcE to cpu6_datapath. It is also
updated at the falling edge but it is the previous value of pcE.

excp_pc is used for exception to get the correct current pc.

pc_auipcE is the correct pcE value for EX stage when handlign
auipc instruction. It does change on the falling edge too.

This is hard to understand and wasteful for pipeline registers.
Need to improve this.