add instructiton jal
add test43_jal test44_beq_jal test45_beq test46_jal2

jal and auipc are different. They both need to add an imm to the
current pc at EX stage. However, jal is a jump instruction. It
assert stallF at IF stage. If stallF is asserted, pcnextF->pcF->
excp_pc will be not be updated.

At EX, auipc instruction uses the pc_auipcE which is excp_pc.
pcE is one step faster than pc_auipcE. Because pcF are
updated in the falling edge, it is always one step faster than the
actual current pc at the next rasing edge. That is why I
need excp_pc to be one step slower than pcF. It is used for
exception handling and the aupic instruction.

For jal, it is a jump. When stallF, excp_pc stops updating, in the
next clock cycle, pcE will be the correct current pc.

branches are the same. Because of stallF, they also use pcE.